Revert "#10 #7 Refactoring, split functionality into separate files, improve readability. Added first (happy flow) unit tests."
This reverts commit 1ff34cb1b14d7932a803a85726a18133c96796c7.
